refactor: extract seedLensDefaults and installLensSkill from ViewDiscovery
